[#9314] PITR: Cleanup sys catalog snapshots
Summary: This diff adds missing logic to delete sys catalog snapshot while removing snapshot object.

Test Plan: ybd --gtest_filter SnapshotScheduleTest.GC

Reviewers: bogdan

Reviewed By: bogdan

Subscribers: skedia, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D12281

+++ src/yb/rocksdb/table/format.cc
@@ -311,7 +311,7 @@ Result<ChecksumData> ComputeChecksum(
           .actual = crc32c::Value(src_data.data(), src_data.size())
       };
     case kxxHash:
-      if (src_data.size() > std::numeric_limits<int>::max()) {
+      if (src_data.size() > static_cast<size_t>(std::numeric_limits<int>::max())) {
         return STATUS_FORMAT(
             Corruption, "Block too large for xxHash ($0 bytes, but must be $1 or smaller)",
             src_data.size(), std::numeric_limits<int>::max());
